requirements.txt
pylib-atle-base-atomic==0.31.0
pylib-atle-base-storage==0.5.0
PyPDF2>=3.0.0
requests>=2.31.0
# OCR stack
pdf2image>=1.17.0
pytesseract>=0.3.10
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.26.0
# GCS support (optional, required for gs://):
google-cloud-storage>=2.8.0